requirements.txt
build
whisper
fastapi
uvicorn
python-dotenv
pipecat-ai[webrtc,daily]
pipecat-ai-small-webrtc-prebuilt
